(roseus_bt) Add separate template generator class
